# openapi/marketplace.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create order (stock reservation, promo, rate-limit)
      description: |
        Roles:
          - USER: allowed
          - ADMIN: allowed
          - SELLER: forbidden
        Business rules:
          - ORDER_LIMIT_EXCEEDED (429)
          - ORDER_HAS_ACTIVE (409)
          - PRODUCT_NOT_FOUND (404)
          - PRODUCT_INACTIVE (409)
          - INSUFFICIENT_STOCK (409, details: list)
          - PROMO_CODE_INVALID (422)
          - PROMO_CODE_MIN_AMOUNT (422)
      x-roles: [USER, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Created
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          description: Product not found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business conflict
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/InsufficientStockError'
              examples:
                order_has_active:
                  value:
                    error_code: ORDER_HAS_ACTIVE
                    message: User already has active order
                    details: null
                product_inactive:
                  value:
                    error_code: PRODUCT_INACTIVE
                    message: Product is inactive
                    details: null
                insufficient_stock:
                  value:
                    error_code: INSUFFICIENT_STOCK
                    message: Insufficient stock
                    details:
                      items:
                        - product_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                          requested: 5
                          available: 2
        '422':
          description: Promo code validation
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    get:
      tags: [Orders]
      operationId: getOrderById
      summary: Get order by id
      description: |
        Roles:
          - USER: only own order (else ORDER_OWNERSHIP_VIOLATION 403)
          - ADMIN: any
      x-roles: [USER, ADMIN]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          description: Access denied or ownership violation
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ownership_violation:
                  value:
                    error_code: ORDER_OWNERSHIP_VIOLATION
                    message: Order belongs to another user
                    details: null
                access_denied:
                  value:
                    error_code: ACCESS_DENIED
                    message: Access denied
                    details: null
        '404':
          description: Order not found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                order_not_found:
                  value:
                    error_code: ORDER_NOT_FOUND
                    message: Order not found
                    details: null

    put:
      tags: [Orders]
      operationId: updateOrder
      summary: Update order items (allowed only in CREATED)
      description: |
        Roles:
          - USER: only own order
          - ADMIN: any
        Rules:
          - only status CREATED (else INVALID_STATE_TRANSITION 409)
          - UPDATE_ORDER rate limit (429)
          - catalog/stock checks (PRODUCT_NOT_FOUND/PRODUCT_INACTIVE/INSUFFICIENT_STOCK)
          - all in one transaction
      x-roles: [USER, ADMIN]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Updated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          description: Access denied or ownership violation
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order or product not found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business conflict
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/InsufficientStockError'
        '429':
          description: Rate limit exceeded
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Promo re-check errors (if promo applied)
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      operationId: cancelOrder
      summary: Cancel order (only CREATED or PAYMENT_PENDING)
      description: |
        Roles:
          - USER: only own order
          - ADMIN: any
        Rules:
          - allowed only from CREATED, PAYMENT_PENDING (else INVALID_STATE_TRANSITION 409)
          - returns stock, decrements promo usage if applied
          - all in one transaction
      x-roles: [USER, ADMIN]
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Canceled
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          description: Access denied or ownership violation
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_state:
                  value:
                    error_code: INVALID_STATE_TRANSITION
                    message: Cancel not allowed from current state
                    details: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    XRequestId:
      description: Request id (UUID) generated per request
      schema:
        type: string
        format: uuid
  parameters:
...
    OrderId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  responses:
    AuthError:
      description: Unauthorized (TOKEN_INVALID / TOKEN_EXPIRED)
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            token_expired:
              value:
                error_code: TOKEN_EXPIRED
                message: Access token expired
                details: null
            token_invalid:
              value:
                error_code: TOKEN_INVALID
                message: Access token invalid
                details: null
    AccessDenied:
      description: Forbidden (ACCESS_DENIED)
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            denied:
              value:
                error_code: ACCESS_DENIED
                message: Access denied
                details: null
    ValidationError:
      description: Validation error (VALIDATION_ERROR)
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            validation_failed:
              value:
                error_code: VALIDATION_ERROR
                message: Validation failed
                details:
                  violations:
                    - field: name
                      rule: minLength
                      message: Must be at least 1 character
                    - field: price
                      rule: minimum
                      message: Must be >= 0.01
  schemas:
    ErrorCode:
      type: string
      enum:
        - PRODUCT_NOT_FOUND
        - PRODUCT_INACTIVE
        - ORDER_NOT_FOUND
        - ORDER_LIMIT_EXCEEDED
        - ORDER_HAS_ACTIVE
        - INVALID_STATE_TRANSITION
        - INSUFFICIENT_STOCK
        - PROMO_CODE_INVALID
        - PROMO_CODE_MIN_AMOUNT
        - ORDER_OWNERSHIP_VIOLATION
        - VALIDATION_ERROR
        - TOKEN_EXPIRED
        - TOKEN_INVALID
        - REFRESH_TOKEN_INVALID
        - ACCESS_DENIED
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          nullable: true
    FieldViolation:
      type: object
      required: [field, rule, message]
      properties:
        field:
          type: string
        rule:
          type: string
        message:
          type: string
    ValidationErrorDetails:
      type: object
      required: [violations]
      properties:
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              $ref: '#/components/schemas/ValidationErrorDetails'
    InsufficientStockItem:
      type: object
      required: [product_id, requested, available]
      properties:
        product_id:
          type: integer
          format: int64
        requested:
          type: integer
          minimum: 1
        available:
          type: integer
          minimum: 0
    InsufficientStockDetails:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InsufficientStockItem'
    InsufficientStockError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              $ref: '#/components/schemas/InsufficientStockDetails'
...
    Money:
      type: number
      format: decimal
      minimum: 0
      multipleOf: 0.01
...
    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]
    OrderItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        promo_code:
          type: string
          nullable: true
          pattern: '^[A-Z0-9_]{4,20}$'
    OrderUpdateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemResponse:
      type: object
      required: [id, product_id, quantity, price_at_order]
      properties:
        id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
        price_at_order:
          allOf:
            - $ref: '#/components/schemas/Money'
          minimum: 0.01
    OrderResponse:
      type: object
      required: [id, user_id, status, total_amount, discount_amount, items, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/OrderStatus'
        promo_code:
          type: string
          nullable: true
        total_amount:
          $ref: '#/components/schemas/Money'
        discount_amount:
          $ref: '#/components/schemas/Money'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...